Add files via upload
This document contains all morphological measurements taken for the oysters from the beginning to the end of the experiment. We used the entire oyster_mm_start and end and the wet_g_start and end weights in our paper. However, more measurements are listed in this document. They include bouyant_g, meaning the weight of the oyster when submerged in water, trying to differentiate the tissue and shell weights. There is the lip_length, which is the very edge of the oyster, where they open and typically experience the most significant growth. Oyster ID represents what they were tagged with on their shells, and Treatment represents which predatory treatment they experienced in the study.