Add files via upload
This document contains water levels from the Lamprey River, which feeds into the Great Bay, where the study was done. Based on the geographic closeness of the two sites, we used this to identify precipitation levels throughout the experiment. This data was collected from the USGS gauge #01073500, pulling measurements for our study duration.