Add files via upload
This is chlorophyll measurements throughout the study. The data was collected by Dr. Elizabeth Harvey's lab at the University of New Hampshire. We did not give oysters any supplementary diet throughout the experiment. We are using the general trends of the chl-a to understand if there was enough food throughout the experiment. We wanted to monitor food abundance to ensure that it did not vary too dramatically, thus leading to profound changes in our oyster's behavioral responses.