Add files via upload
This document contains information from a YSI probe that was used consistently to monitor the parameters of the water. This document includes when cages were cleaned, meaning the mesh covering the tanks to stop crabs from escaping was washed, and the manifold was flushed to remove buildup. Additionally, some days have NA, which means that YSI measurements were not obtained at this time. Additionally, information from the saltwater master test kit for nitrates, nitrites, and ammonia measurements is included. There are also notes regarding oysters and crabs, equipment technical difficulties, and facilities maintenance through the experiment for reference when conducting the analysis.