Add files via upload
Chlorophyll-a measurements collected by the Harvey lab at the University of New Hampshire. These water samples were collected from the pier outside of the facility where the experiment was conducted. The experiment used a flow-through system, thus taking in the water from Great Bay directly outside the facility near where these water samples were collected. The chl_a column was used in the manuscript to identify general phytoplankton estimate trends in the bay throughout the experiment. The chl_a column uses fluorescence readings and calibrates the raw data by accounting for fluorescence readings by using the original RFU reading and the RFU readings after two drops of HCL are added to the sample. This accounts for biological materials in the sample that are living and gives a more accurate measurement of the phytoplankton abundance in the samples.